+++ CHANGELOG.md
@@ -2,6 +2,12 @@ Changelog
 =========
 
 
+0.0.2 (2021-05-20)
+------------------
+- Fix verification of Consul connection when URL is invalid
+- Fix Consul init when environment has invalid variables
+
+
 0.0.1 (2021-05-15)
 ------------------
 - First version